# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: BigFin API
  version: 0.1.0
  description: |
    Multi-tenant loan administration API powering GreenBucks and other operator apps.

    ## Key Concepts
    - **Tenant**: Operator app (e.g., GreenBucks) with isolated data
    - **EndCustomer**: Borrower or lender on the platform
    - **LoanOffer**: Proposed loan terms from lender to borrower
    - **LoanContract**: Accepted offer with repayment schedule
    - **PaymentIntent**: Idempotent payment request (disbursement or repayment)

    ## Idempotency
    All POST endpoints that create resources or initiate payments accept an `Idempotency-Key` header.
    Requests with the same key within 24 hours return the original response.

    ## Error Codes
    See the Error Taxonomy section at the bottom of this spec.
  contact:
    name: BigFin API Support
  license:
    name: Proprietary
servers:
  - url: https://api.bigfin.io/v1
    description: Production
  - url: https://sandbox.api.bigfin.io/v1
    description: Sandbox
security:
  - BearerAuth: []
...
paths:
...
  /loan-contracts/{contract_id}/repayments:
    parameters:
      - $ref: '#/components/parameters/ContractId'
    post:
      tags: [Repayments]
      summary: Initiate repayment
      description: |
        Borrower initiates a payment. Amount is applied according to payment waterfall policy.
        Can be scheduled or immediate.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RepaymentRequest'
      responses:
        '202':
          description: Repayment initiated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Repayment'

    get:
      tags: [Repayments]
      summary: List repayments
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [scheduled, initiated, pending, completed, failed, returned]
        - $ref: '#/components/parameters/PageCursor'
        - $ref: '#/components/parameters/PageLimit'
      responses:
        '200':
          description: Repayment list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RepaymentList'
  /repayments/{repayment_id}:
    parameters:
      - $ref: '#/components/parameters/RepaymentId'
    get:
      tags: [Repayments]
      summary: Get repayment status
      responses:
        '200':
          description: Repayment details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Repayment'
  /repayments/{repayment_id}/cancel:
    parameters:
      - $ref: '#/components/parameters/RepaymentId'
    post:
      tags: [Repayments]
      summary: Cancel scheduled repayment
      description: Only scheduled (not yet initiated) repayments can be cancelled
      responses:
        '200':
          description: Repayment cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Repayment'
        '409':
          description: Repayment already initiated
          $ref: '#/components/responses/Conflict'

  # ============ DOCUMENTS ============
...
  /webhooks:
    post:
      tags: [Webhooks]
      summary: Register webhook endpoint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWebhookRequest'
      responses:
        '201':
          description: Webhook registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Webhook'

    get:
      tags: [Webhooks]
      summary: List webhook endpoints
      responses:
        '200':
          description: Webhook list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookList'
  /webhooks/{webhook_id}:
    parameters:
      - name: webhook_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    delete:
      tags: [Webhooks]
      summary: Delete webhook endpoint
      responses:
        '204':
          description: Webhook deleted

  # ============ OPERATOR ENDPOINTS ============
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
        format: uuid
      description: Unique key for idempotent requests. Same key within 24h returns cached response.
    PageCursor:
      name: cursor
      in: query
      schema:
        type: string
      description: Pagination cursor from previous response
    PageLimit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    ContractId:
      name: contract_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    RepaymentId:
      name: repayment_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    FundsAvailability:
      type: object
      properties:
        state:
          type: string
          enum: [initiated, pending, received, held, available, failed]
        available_at:
          type: string
          format: date-time
        hold_reason:
          type: string
        hold_expires_at:
          type: string
          format: date-time

    # ===== Repayments =====
    RepaymentRequest:
      type: object
      required: [amount_cents]
      properties:
        amount_cents:
          type: integer
          minimum: 1
        funding_instrument_id:
          type: string
          format: uuid
        scheduled_date:
          type: string
          format: date
          description: Schedule for future date (omit for immediate)
        is_payoff:
          type: boolean
          description: If true, amount must equal payoff amount
    Repayment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        contract_id:
          type: string
          format: uuid
        amount_cents:
          type: integer
        status:
          type: string
          enum: [scheduled, initiated, pending, completed, failed, returned]
        rail:
          type: string
        funding_instrument_id:
          type: string
          format: uuid
        application:
          $ref: '#/components/schemas/PaymentApplication'
        availability:
          $ref: '#/components/schemas/FundsAvailability'
        scheduled_date:
          type: string
          format: date
        initiated_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        failure_reason:
          type: string
    RepaymentList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Repayment'
        next_cursor:
          type: string
        has_more:
          type: boolean
    PaymentApplication:
      type: object
      description: How the payment was applied per waterfall policy
      properties:
        fees_cents:
          type: integer
        interest_cents:
          type: integer
        principal_cents:
          type: integer
        breakdown:
          type: array
          items:
            type: object
            properties:
              bucket:
                type: string
              amount_cents:
                type: integer

    # ===== Documents =====
...
    CreateWebhookRequest:
      type: object
      required: [url, events]
      properties:
        url:
          type: string
          format: uri
        events:
          type: array
          items:
            type: string
            enum:
              - loan_offer.sent
              - loan_offer.accepted
              - loan_offer.rejected
              - loan_offer.expired
              - loan_contract.created
              - loan_contract.status_changed
              - disbursement.initiated
              - disbursement.completed
              - disbursement.failed
              - repayment.initiated
              - repayment.completed
              - repayment.failed
              - repayment.returned
              - funds.available
              - prefund.deposit_completed
              - prefund.withdrawal_completed
              - kyc.status_changed
        secret:
          type: string
          description: Shared secret for HMAC signature (auto-generated if not provided)
    Webhook:
      type: object
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
          format: uri
        events:
          type: array
          items:
            type: string
        secret:
          type: string
          description: Only returned on creation
        status:
          type: string
          enum: [active, disabled]
        created_at:
          type: string
          format: date-time
    WebhookList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Webhook'
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error code
        message:
          type: string
          description: Human-readable message
        details:
          type: object
          description: Additional context
        request_id:
          type: string
          description: Request ID for support
  responses:
...
    Conflict:
      description: Conflict with current state
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: "invalid_state"
            message: "Offer has already been accepted"

# ===== Error Taxonomy =====
#
# Error codes follow the pattern: {category}_{specific}
#
# Authentication/Authorization:
#   - unauthorized: Missing or invalid token
#   - forbidden: Valid token but insufficient permissions
#   - step_up_required: Sensitive action requires re-authentication
#
# Validation:
#   - invalid_request: Malformed request body
#   - invalid_parameter: Specific parameter validation failed
#   - terms_out_of_policy: Loan terms violate product policy
#
# State:
#   - invalid_state: Action not allowed in current state
#   - already_exists: Duplicate resource (use idempotency key)
#   - not_found: Resource doesn't exist
#
# Payment:
#   - insufficient_funds: Not enough balance
#   - instrument_invalid: Funding instrument can't be used
#   - payment_failed: Payment provider rejected
#   - payment_returned: Payment was returned/reversed
#
# Limits:
#   - limit_exceeded: Transaction or balance limit hit
#   - rate_limited: Too many requests
#
# System:
#   - internal_error: Unexpected server error
#   - service_unavailable: Temporary outage
#   - provider_error: Payment provider issue